Refactor file structure by moving HomePage and DashboardPage components to their respective directories. Update import paths accordingly. Remove unused HomePage component.
