+++ Docs/journaling/08-15-25.md
@@ -1,0 +1,41 @@
+# Journal - 15-08-25
+
+## Réintégration manuelle du Dashboard
+
+- Réintégration manuelle du nouveau Dashboard dans le projet
+- Question sur la nécessité du fichier index.ts dans le dossier dashboard
+- Analyse de la structure d'importation actuelle
+
+## Actions effectuées
+
+- Suppression du fichier `index.ts` superflu dans `client/src/components/dashboard/`
+- Mise à jour des imports dans `DashboardPage.tsx` pour des imports directs
+- Simplification de la structure d'importation
+
+## Réorganisation de la structure des pages
+
+- Création du dossier `client/src/pages/dashboard/` et déplacement de `DashboardPage.tsx`
+- Création du dossier `client/src/pages/home/` et déplacement de `HomePage.tsx`
+- Mise à jour des imports dans `App.tsx` pour refléter la nouvelle structure
+- Suppression des anciens fichiers isolés
+
+## Correction des chemins d'import
+
+- Correction des imports dans `HomePage.tsx` : `../` → `../../` (contexts, hooks, types, components, assets)
+- Correction des imports dans `DashboardPage.tsx` : `../` → `../../` (contexts, hooks, components)
+- Résolution de l'erreur Vite "Failed to resolve import"
+
+## Résultat
+
+- Structure plus simple et directe
+- Moins de fichiers à maintenir
+- Imports plus explicites et clairs
+- **Organisation cohérente** : toutes les pages sont maintenant dans des dossiers organisés
+- Structure plus maintenable et professionnelle
+- **Erreurs d'import résolues** ✅
+
+## Question sur l'organisation
+
+- DashboardPage.tsx et HomePage.tsx sont dans le dossier pages
+- Tous les autres composants sont dans des dossiers organisés
+- Possibilité de créer un dossier dashboard pour DashboardPage.tsx ?
