+++ Docs/journaling/12-15-25.md
@@ -1,0 +1,19 @@
+# Journal - 15-12-25
+
+## Réintégration manuelle du Dashboard
+
+- Réintégration manuelle du nouveau Dashboard dans le projet
+- Question sur la nécessité du fichier index.ts dans le dossier dashboard
+- Analyse de la structure d'importation actuelle
+
+## Actions effectuées
+
+- Suppression du fichier `index.ts` superflu dans `client/src/components/dashboard/`
+- Mise à jour des imports dans `DashboardPage.tsx` pour des imports directs
+- Simplification de la structure d'importation
+
+## Résultat
+
+- Structure plus simple et directe
+- Moins de fichiers à maintenir
+- Imports plus explicites et clairs
